requirements.txt
asgiref==3.2.3
Django==3.0.4
djangorestframework==3.11.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
python-dotenv==0.12.0
pytz==2019.3
sqlparse==0.3.1
